+++ src/ArrayyMeta.php
@@ -47,6 +47,19 @@ final class ArrayyMeta
             }
         }
 
+        /** @noinspection PhpAssignmentInConditionInspection */
+        while ($reflector = $reflector->getParentClass()) {
+            $docComment = $reflector->getDocComment();
+            if ($docComment) {
+                $docblock = $factory->create($docComment);
+                /** @var \phpDocumentor\Reflection\DocBlock\Tags\Property $tag */
+                foreach ($docblock->getTagsByName('property') as $tag) {
+                    $PropertyName = $tag->getVariableName();
+                    $this->{$PropertyName} = $PropertyName;
+                }
+            }
+        }
+
         $STATIC_CACHE[$cacheKey] = $this;
 
         return $this;