+++ tests/BigCityData.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Arrayy\tests;
+
+/**
+ * @property string $extra_info
+ */
+class BigCityData extends CityData
+{
+}